Restored user load behavior, except using always_200=true
