+++ src/index.html
@@ -25,17 +25,20 @@
         <link rel="prefetch" href="<%= process.env.API_ORIGIN %>/apps/cms/api/books/?format=json">
         <link rel="prefetch" href="<%= process.env.API_ORIGIN %>/apps/cms/api/footer/?format=json">
         <script>
-          window._OX_USER_PROMISE ||= fetch('/accounts/api/user?always_200=true', { credentials: 'include' }).then(
-            (response) => {
-              if (response.ok) {
-                return response.json().catch((error) => {
-                  throw new Error(`Failed to parse user JSON: ${error.message}`);
-                });
-              } else {
-                throw new Error(`Failed to load user: HTTP ${response.status} status code`);
+          // This code is shared with the accounts-model
+          window._OX_USER_PROMISE ||= fetch('/accounts/api/user?always_200=true', {credentials: 'include'}).then(
+              (response) => response.json().catch(
+                  (err) => {
+                      console.warn('No JSON in Accounts response');
+                      return {err};
+                  }
+              ),
+              (err) => {
+                  console.warn('"Error fetching user info"');
+                  return {err};
               }
-            }
           );
+
           window._OX_USER_PROMISE.then((user) => {
             const updateUserConsentPreferences = (consentPreferences) => {
               if (user.id) {
