--- src/typings/global.d.ts
@@ -1,8 +1,0 @@
-import type { AccountsUserModel } from './accounts';
-
-declare global {
-    interface Window {
-        _OX_USER_PROMISE?: Promise<AccountsUserModel>;
-        dataLayer?: object[];
-    }
-}
